mavenproject2/src/test/java/com/mycompany/mavenproject2/StudentTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.mycompany.mavenproject2;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author Tayyab
 */
public class StudentTest {
    
    public StudentTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }
    
    @Before
    public void setUp() {
    }
    
    @After
    public void tearDown() {
    }

    /**
     * Test of setanswer method, of class Student.
     */
    @Test
    public void testSetanswer() {
        System.out.println("setanswer");
        Integer answer = null;
        Student instance = new Student();
        instance.setanswer(answer);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getanswer method, of class Student.
     */
    @Test
    public void testGetanswer() {
        System.out.println("getanswer");
        Student instance = new Student();
        Integer expResult = null;
        Integer result = instance.getanswer();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of setquestion method, of class Student.
     */
    @Test
    public void testSetquestion() {
        System.out.println("setquestion");
        String question = "";
        Student instance = new Student();
        instance.setquestion(question);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getquestion method, of class Student.
     */
    @Test
    public void testGetquestion() {
        System.out.println("getquestion");
        Student instance = new Student();
        String expResult = "";
        String result = instance.getquestion();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of setId method, of class Student.
     */
    @Test
    public void testSetId() {
        System.out.println("setId");
        Integer id = null;
        Student instance = new Student();
        instance.setId(id);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getId method, of class Student.
     */
    @Test
    public void testGetId() {
        System.out.println("getId");
        Student instance = new Student();
        Integer expResult = null;
        Integer result = instance.getId();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }
    
}
